+++ src/generator/gen_json.rs
@@ -21,7 +21,7 @@ impl GeneratorJson {
     }
 
     fn desc_to_string(&mut self, desc: &Description) -> String {
-        desc.get().replace("\n","\\n")
+        desc.get(self.is_public()).replace("\n","\\n")
     }
 
 }