+++ src/rifgen/description.rs
@@ -2,51 +2,79 @@ use crate::parser::{logic_expr, parser_expr::{parse_expr, ParamValues}};
 
 /// Thin wrapper around string to ease update and get short/long descripion
 #[derive(Clone, Debug, PartialEq, Default)]
-pub struct Description(String);
+pub struct Description {
+    public: String,
+    private: String,
+}
 
 impl Description {
-    pub fn updt(&mut self, desc: &str) {
-        if !self.0.is_empty() {
-            self.0.push('\n');
+
+    pub fn updt(&mut self, desc: &str, is_private: bool) {
+        let s = if is_private {&mut self.private} else {&mut self.public};
+        if !s.is_empty() {
+            s.push('\n');
         }
         let mut line_it = desc.split("\\n").peekable();
         while let Some(line) = line_it.next() {
-            self.0.push_str(line);
+            s.push_str(line);
             if line_it.peek().is_some() {
-                self.0.push('\n');
+                s.push('\n');
             }
         }
     }
 
-    pub fn get(&self) -> &str {
-        &self.0
+    pub fn get(&self, is_public: bool) -> String {
+        let mut desc = String::with_capacity(self.len(is_public));
+        desc.push_str(&self.public);
+        if !is_public && !self.private.is_empty(){
+            desc.push('\n');
+            desc.push_str(&self.private);
+        }
+        desc
     }
 
-    pub fn get_short(&self) -> &str {
-        if let Some(end) = self.0.find('\n') {
-            &self.0[..end]
+    pub fn get_short(&self, is_public: bool) -> String {
+        let s = if self.public.is_empty() && !is_public {
+            &self.private
         } else {
-            &self.0
+            &self.public
+        };
+        if let Some(end) = s.find('\n') {
+            s[..end].to_owned()
+        } else {
+            s.to_owned()
         }
     }
 
-    pub fn get_split(&self) -> (&str, Option<&str>) {
-    	match self.0.find('\n') {
-    		Some(mid) if mid < self.0.len()-1 => (&self.0[..mid], Some(&self.0[mid+1..self.0.len()])),
-    		Some(end)  => (&self.0[..end], None),
-    		None => (&self.0, None),
+    pub fn get_split(&self, is_public: bool) -> (String, Option<String>) {
+        let d = self.get(is_public);
+    	match d.find('\n') {
+    		Some(mid) if mid < d.len()-1 => (d[..mid].to_owned(), Some(d[mid+1..d.len()].to_owned())),
+    		Some(end)  => (d[..end].to_owned(), None),
+    		None => (d, None),
     	}
     }
 
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
+    /// Check if the description is empty
+    pub fn is_empty(&self, is_public: bool) -> bool {
+        self.public.is_empty() && (is_public || self.private.is_empty())
+    }
+
+    /// Check if the description is empty
+    pub fn len(&self, is_public: bool) -> usize {
+        self.public.len() + (if is_public {0} else {self.private.len()})
     }
 
     pub fn interpolate(&self, idx: u16) -> Description {
-        // if self.0.starts_with("Gain and") {println!("{}",self.0)};
-        let mut desc = String::with_capacity(self.0.len());
+        let public = Self::interpolate_str(&self.public, idx);
+        let private = Self::interpolate_str(&self.private, idx);
+        Description {public, private}
+    }
+
+    fn interpolate_str(s: &str, idx: u16) -> String {
+        let mut desc = String::with_capacity(s.len());
         let params = ParamValues::new_with_idx(idx as isize);
-        let mut parts = self.0.split('$');
+        let mut parts = s.split('$');
         desc.push_str(parts.next().unwrap_or(""));
         for mut s in parts {
             // Variable $i replaced by index
@@ -65,31 +93,33 @@ impl Description {
                 desc.push_str(s);
             }
         }
-        Description(desc)
+        desc
     }
 
     // Create a description with $f replace by a format string (for example u8.0 or s5.3)
     pub fn with_format(&self, format: &str) -> Description {
-        let desc = self.0.replace("$f", format);
-        Description(desc)
+        let public = self.public.replace("$f", format);
+        let private = self.private.replace("$f", format);
+        Description{public, private}
     }
 
     // Remove the $ special character from description:
     //  used when we do not want the interpolated version (for register base description typically)
     pub fn no_dollar(&self) -> Description {
-        let desc = self.0.replace('$', "");
-        Description(desc)
+        let public = self.public.replace('$', "");
+        let private = self.private.replace('$', "");
+        Description{public, private}
     }
 }
 
 impl From<String> for Description {
-    fn from(d: String) -> Description {
-        Description(d)
+    fn from(s: String) -> Description {
+        Description{public: s, private: "".to_owned()}
     }
 }
 
 impl From<&str> for Description {
-    fn from(v: &str) -> Description {
-        Description(v.to_owned())
+    fn from(s: &str) -> Description {
+        Description{public: s.to_owned(), private: "".to_owned()}
     }
 }